fix Controller, Service

+++ src/main/java/com/devsuperior/DSClient/service/ClientService.java
@@ -1,25 +1,34 @@
 package com.devsuperior.DSClient.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.devsuperior.DSClient.dto.ClientDTO;
 import com.devsuperior.DSClient.entities.Client;
 import com.devsuperior.DSClient.repositores.ClientRepository;
+import com.devsuperior.DSClient.service.exception.DatabaseException;
+import com.devsuperior.DSClient.service.exception.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ClientService {
+
     @Autowired
     private ClientRepository repository;
     
     @Transactional(readOnly = true)
     public ClientDTO findById(Long id){
-        Client client = repository.findById(id).get();
+        Client client = repository.findById(id).orElseThrow(
+            () -> new ResourceNotFoundException("Recurso não encontrado"));
         return new ClientDTO(client);
     }
+
     @Transactional(readOnly = true)
     public Page<ClientDTO> findAll(Pageable pageable){
         Page<Client> result = repository.findAll(pageable);
@@ -37,18 +46,30 @@ public class ClientService {
 
     @Transactional
     public ClientDTO update(Long id, ClientDTO dto){
-
+    try{
         Client entity = repository.getReferenceById(id);
         copyDtoEntity(dto, entity);
         entity = repository.save(entity);
         return new ClientDTO(entity);
-
+       }
+       catch (EntityNotFoundException e){
+        throw new ResourceNotFoundException("Recurso não encontrado");
+       }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id){
-        repository.deleteById(id);
+        if(!repository.existsById(id)){
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try{
+            repository.deleteById(id);            
+        }
+        catch (DataIntegrityViolationException e){
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
+
 
     private void copyDtoEntity(ClientDTO dto, Client entity) {
 
@@ -58,5 +79,7 @@ public class ClientService {
         entity.setBirthDate(dto.getBirthDate());
         entity.setChildren(dto.getChildren());
     }
+
+
 
 }
